// www2/src/app/models/target-inbound-message.model.ts
import { Bumper } from './bumper.model';

export const OUT_COMPUTER_CONNECTED = 0x80;
export const OUT_COMPUTER_CALIBRATION_STARTED = 0x81;
export const OUT_COMPUTER_CALIBRATION_FINISHED = 0x82;
export const OUT_COMPUTER_BUMPER_HIT = 0x84;
export const OUT_COMPUTER_CONTROLLER_STATE = 0x88;

export enum OutCode {
  OUT_COMPUTER_CONNECTED = 0x80,
  OUT_COMPUTER_CALIBRATION_STARTED = 0x81,
  OUT_COMPUTER_CALIBRATION_FINISHED = 0x82,
  OUT_COMPUTER_BUMPER_HIT = 0x84,
  OUT_COMPUTER_CONTROLLER_STATE = 0x88
}

export interface TargetInboundMessage {
  code: OutCode;
  state?: Bumper[];
}